gh-93649: Split _testcapimodule.c into smaller files
* Move many functions from _testcapimodule.c into more specific files
  in Modules/_testcapi/.

* Add files:

  * Modules/_testcapi/frame.c
  * Modules/_testcapi/function.c
  * Modules/_testcapi/type.c

* In moved code:

  * Replace get_testerror() with PyExc_AssertionError.
  * Replace raiseTestError() with
    PyErr_Format(PyExc_AssertionError, ...).
