+++ Modules/_testcapi/parts.h
@@ -62,5 +62,8 @@ int _PyTestCapi_Init_Monitoring(PyObject *module);
 int _PyTestCapi_Init_Object(PyObject *module);
 int _PyTestCapi_Init_Config(PyObject *mod);
 int _PyTestCapi_Init_Import(PyObject *mod);
+int _PyTestCapi_Init_Type(PyObject *mod);
+int _PyTestCapi_Init_Function(PyObject *mod);
+int _PyTestCapi_Init_Frame(PyObject *mod);
 
 #endif // Py_TESTCAPI_PARTS_H
